requirements.txt
argcomplete==1.12.3
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
commitizen==2.29.4
decli==0.5.2
Django==3.2
djangorestframework==3.12.4
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
sqlparse==0.4.2
termcolor==1.1.0
tomlkit==0.11.1
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5

psycopg2
celery==4.4.1
django-celery-beat==2.0.0
django-celery-results==1.2.0
flower==0.9.5